Discontinue support for the PHP versions 5 and 7
It is becoming increasingly difficult to support three versions of the
code (php5, php7 and php8). The latest changes in PHP 8.3.10 complicated
the build scripts by the requirement to have the header files in the
extension's root directory. For example, when compiling in the PHP
source, there must be a `php_ev.h` (or the name of the main header) in
the `ext/ev/` directory. For more details please refer to the following
files in the PHP source:
- build/php.m4 (in particular, the `PHP_NEW_EXTENSION` macro)
- build/genif.sh (`header_list="$header_list $ext_dir/*.h*"`)
- build/print_include.awk

Also, even the official PHP releases older than 8.1 are currently
unsupported. See https://www.php.net/supported-versions.php.

This should also fix the issue #51:
https://bitbucket.org/osmanov/pecl-ev/issues/51/building-statically-into-php-fails-on-php